+++ README.md
@@ -1,3 +1,6 @@
+## Лабораторная работа 5
+# Выполнил Тимофеев А.А.
+
 # Объектно-ориентированное программирование
 
 ### Внешняя декомпозиция
