+++ student_info.cpp
@@ -1,0 +1,35 @@
+#include "Student_info.h"
+#include <iostream>
+#include <string>
+#include <vector>
+
+using std::istream;
+using std::string;
+using std::vector;
+
+//==============================================================================
+istream& read_hw (istream& in, vector<double>& hw) {
+	if (in) {
+		hw.clear();
+		
+		for (double x; in >> x;) hw.push_back(x);
+
+		in.clear();
+	}
+	return in;
+}
+
+
+//==============================================================================
+istream& read (istream& is, Student_info& s) {
+	is >> s.name >> s.midterm >> s.final;
+	read_hw(is, s.homework);
+
+	return is;
+}
+
+
+//==============================================================================
+bool compare (const Student_info& x, const Student_info& y) {
+	return x.name < y.name;
+}
